\name{DMwR-package}
\alias{DMwR-package}
\alias{DMwR}
\docType{package}
\title{
Functions and data for the book "Data Mining with R"
}
\description{
This package includes functions and data accompanying the book
"Data Mining with R, learning with case studies" by Luis Torgo,
published by CRC Press (ISBN: 9781439810187)
}
\author{
Luis Torgo

Maintainer: Luis Torgo \email{ltorgo@dcc.fc.up.pt}
}
\references{
Torgo, L. (2010) \emph{Data Mining using R: learning with case studies},
CRC Press (ISBN: 9781439810187). \url{http://www.dcc.fc.up.pt/~ltorgo/DataMiningWithR}  
}
\keyword{ package }
