\name{expSettings-class}
\Rdversion{1.1}
\docType{class}
\alias{expSettings}
\alias{expSettings-class}

\title{Class "expSettings"  }
\description{	 This is a class union formed by the classes cvSettings,
  mcSettings, hldSettings, loocvSettings and bootSettings}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Methods}{
No methods defined with class "expSettings" in the signature.
}
\references{ Torgo, L. (2010) \emph{Data Mining using R: learning with case studies},
CRC Press (ISBN: 9781439810187). \url{http://www.dcc.fc.up.pt/~ltorgo/DataMiningWithR}  
}
\author{ Luis Torgo \email{ltorgo@dcc.fc.up.pt} }
\seealso{
  \code{\linkS4class{cvSettings}},   \code{\linkS4class{mcSettings}},
  \code{\linkS4class{loocvSettings}}, \code{\linkS4class{hldSettings}},\code{\linkS4class{bootSettings}}
}
\examples{
showClass("expSettings")
}
\keyword{classes}
