\name{DMwR-defunct}
%------ NOTE:  ../R/DMwR-defunct.R   must be synchronized with this!
\alias{DMwR-defunct}
\alias{slidingWindowTesting}
\alias{growingWindowTesting}

\title{Defunct Functions in Package \pkg{DMwR}}
\description{
  The functions or variables listed here are no longer part of package \pkg{DMwR} .
}
\usage{
slidingWindowTesting(learner,form,train,test,relearn.step=1,verbose=T)

growingWindowTest(learner,form,train,test,relearn.step=1,verbose=T)

}
\seealso{
  \code{\link{Deprecated}}, \code{\link{Defunct}}
}
\keyword{internal}
