\name{DMwR-defunct}
%------ NOTE:  ../R/DMwR-defunct.R   must be synchronized with this!
\alias{DMwR-defunct}
\alias{classWF}
\alias{regrWF}
\alias{slideRegrWF}
\alias{growRegrWF}
\alias{tsRegrWF}
\alias{slideClassWF}
\alias{growClassWF}
\alias{tsClassWF}

\title{Defunct Functions in Package \pkg{DMwR}}
\description{
  The functions or variables listed here are no longer part of package \pkg{DMwR} .
}
\usage{

classWF(form,train,test,learner,eval,...)

regrWF(form,train,test,learner,eval,simpl,...)

slideRegrWF(...)

growRegrWF(...)

tsRegrWF(form,train,test,type,learner,eval,simpl,relearn.step,verbose,...)

slideClassWF(...)

growClassWF(...)

tsClassWF(form,train,test,type,learner,eval,relearn.step,verbose,...)
}
\seealso{
  \code{\link{Deprecated}}, \code{\link{Defunct}}
}
\keyword{internal}
